requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
psycopg2-binary
jinja2
aiofiles
deep-translator
cohere
mysql-connector-python>=8.0.33
python-dotenv>=1.1.1
python-multipart
